+++ src/format.rs
@@ -7,7 +7,7 @@ use core::fmt::{self, Write as _};
 
 #[cfg(feature = "alloc")]
 use alloc::collections::TryReserveError;
-#[cfg(feature = "alloc")]
+#[cfg(all(feature = "alloc", not(feature = "std")))]
 use alloc::string::String;
 
 /// Output buffer capacity overflow error.