+++ src/lib.rs
@@ -134,7 +134,9 @@
 //! [`RiReferenceStr::resolve_against()`]: `types::RiReferenceStr::resolve_against`
 //! [`RiRelativeStr::resolve_against()`]: `types::RiRelativeStr::resolve_against`
 #![warn(missing_docs)]
+#![warn(unsafe_op_in_unsafe_fn)]
 #![warn(clippy::missing_docs_in_private_items)]
+#![warn(clippy::undocumented_unsafe_blocks)]
 #![cfg_attr(not(feature = "std"), no_std)]
 #![cfg_attr(docsrs, feature(doc_cfg))]